# algohive/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Catalog:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
    type: object
  models.Competition:
    properties:
      api_environment:
        $ref: '#/definitions/models.Catalog'
      api_environment_id:
        type: string
      api_theme:
        type: string
      description:
        type: string
      finished:
        type: boolean
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      show:
        type: boolean
      title:
        type: string
      tries:
        items:
          $ref: '#/definitions/models.Try'
        type: array
    type: object
  models.Group:
    properties:
      competitions:
        items:
          $ref: '#/definitions/models.Competition'
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scope_id:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Role:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        type: integer
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Scope:
    properties:
      catalogs:
        items:
          $ref: '#/definitions/models.Catalog'
        type: array
      description:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  models.Try:
    properties:
      attempts:
        type: integer
      competition_id:
        type: string
      end_time:
        type: string
      id:
        type: string
      puzzle_id:
        type: string
      puzzle_index:
        type: integer
      puzzle_lvl:
        type: string
      score:
        type: number
      start_time:
        type: string
      step:
        type: integer
      user_id:
        type: string
    type: object
  models.User:
    properties:
      blocked:
        type: boolean
      email:
        type: string
      firstname:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      last_connected:
        type: string
      lastname:
        type: string
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  users.PasswordUpdate:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
info:
  contact:
    email: ericphlpp@proton.me
    name: AlgoHive Support
  description: This is the API documentation for the AlgoHive API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger AlgoHive API
  version: 1.0.0
paths:
...
  /scopes/{scope_id}/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Detach the scope from a role, only accessible to users with the
        SCOPES permission
      parameters:
      - description: Scope ID
        in: path
        name: scope_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scope'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Detach the scope from a role
      tags:
      - Scopes
    post:
      consumes:
      - application/json
      description: Attach the scope to a role, only accessible to users with the SCOPES
        permission
      parameters:
      - description: Scope ID
        in: path
        name: scope_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scope'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Attach the scope to a role
      tags:
      - Scopes
  /scopes/roles:
    get:
      consumes:
      - application/json
      description: Get all the scopes that a role has access to, only accessible to
        users with the SCOPES permission
      parameters:
      - collectionFormat: csv
        description: Roles IDs
        in: query
        items:
          type: string
        name: roles
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Scope'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all the scopes that a role has access to
      tags:
      - Scopes
  /scopes/user:
    get:
      consumes:
      - application/json
      description: Get all scopes that the user has access to (based on roles) if
        the user has the SCOPES permission we return all scopes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Scope'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all scopes that the user has access to
      tags:
      - Scopes
...
  /user/profile:
    get:
      description: Get the profile information of the authenticated user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get User Profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile information of the authenticated user
      parameters:
      - description: User Profile
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update User Profile
      tags:
      - Users
  /user/profile/password:
    put:
      consumes:
      - application/json
      description: Update the password of the current user
      parameters:
      - description: Password Update
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/users.PasswordUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update User Password
      tags:
      - Users
  /user/resetpass:
    put:
      consumes:
      - application/json
      description: Reset the password of the target user
      parameters:
      - description: User Profile
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Reset Target User Password
      tags:
      - Users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"